feat(migrations): Add stored procedures for product limits and membership management
- Migration 036: Create stored procedure `validar_limite_productos` to validate if a local can add more products based on their membership.
- Migration 037: Update stored procedure `obtener_todas_membresias` to include limit fields for memberships.
- Migration 038: Fix stored procedure `obtener_suscripcion_activa` to correctly retrieve active subscriptions and auto-renew free plans.
- Migration 039: Correct stored procedure `obtener_suscripcion_activa` to remove non-existent fields and include membership limit fields.
- Migration 040: Fix stored procedure `listar_galeria_pendientes` to include missing columns for gallery review.
- Migration 041: Correct stored procedure `obtener_limites_membresia` to cast product limits from VARCHAR to INT for comparisons.
- Migration 042: Fix stored procedure `obtener_limites_membresia` to use correct column names in `oferta_flash`.
- Migration 043: Finalize corrections in `obtener_limites_membresia` to use the correct tables and columns for product counts.
- Migration 044: Change `DOCUMENTO_IDENTIDAD` from INT to BIGINT in `persona` table to accommodate longer identity documents.
- Migration 045: Update `logs_sistema` table to include LOGIN and LOGOUT actions in `TIPO_ACCION` and modify the logging procedure accordingly.

+++ BusinessLogic/ManejadorPersonas.cs
@@ -44,7 +44,7 @@ namespace BusinessLogic
                     Apellido = row["APELLIDOS"].ToString(),
                     Correo = row["CORREO"].ToString(),
                     Telefono = row["TELEFONO"].ToString(),
-                    Documento = Convert.ToInt32(row["DOCUMENTO_IDENTIDAD"].ToString()),
+                    Documento = Convert.ToInt64(row["DOCUMENTO_IDENTIDAD"].ToString()),
                     Estado = row["ESTADO"].ToString()
                 });
             }
